potentially wrong result when compare float with double: #142

+++ Demo/YYTextDemo/NSString+YYAdd.h
@@ -119,7 +119,7 @@
  */
 - (BOOL)containsEmoji;
 
-- (BOOL)containsEmojiForSystemVersion:(float)systemVersion;
+- (BOOL)containsEmojiForSystemVersion:(double)systemVersion;
 
 
 #pragma mark - Utilities
